[pxe_udp] Place an upper bound on number of queued packets
Currently, when pxe_udp is active, it continues to enqueue packets with
no upper bound, which could lead to consuming all available memory if
the queue is not consumed fast enough (or at all) while under heavy
traffic. Once all memory is consumed by this queue, drivers are unable
to allocate rx buffers for new packets, rendering all other sockets
useless.

Place an upper bound on the length of this queue to limit the amount of
memory it can consume.

Signed-off-by: Tal Shorer <shorer@amazon.com>
Reviewed-by: Itamar Katz <itamark@amazon.com>