[efi] Rename "driver" to "target" in preparation for two-phase veto mechanism
Forcible uninstallation of handles installed by a vetoed driver can be
done only on the basis of an image handle, since there is no way
within the UEFI driver model to identify the driver that installed a
protocol on a handle.  A single image may provide multiple drivers.

This will require a two-phase mechanism in which we first perform the
DisconnectController() calls, uninstall the driver binding protocols,
and close any stray handles associated with the driver handle.  Once
all driver binding handles have been processed in this way, a second
pass can close and uninstall any stray handles associated with the
image handle.

Prepare for a two-phase veto mechanism by renaming the "driver" handle
to a "target" handle (which may be either the driver handle or the
image handle).

Signed-off-by: Michael Brown <mcb30@ipxe.org>